Support nil pointers in JSON codec
When a nil value of a struct type was used with the JSON codec (e.g.
`(*MyType)(nil)`) it would cause an error in the round-trip check for
equality:

```
value differs after marshalling and unmarshalling: values are not equal (-want +got):
          (*pkg.MyType)(
        -       nil,
        +       &{},
          )
```

The reason was that unmarshalling functions received a pointer to the
destination structure instead of a pointer to a pointer which could be
set to nil.